class based views

+++ store/views-fn-based.py
@@ -5,40 +5,39 @@ from django.shortcuts import render, get_object_or_404
 from rest_framework.decorators import api_view
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.views import APIView
 
 
-class ProductList(APIView):
-    def get(self, request):
+@api_view(['GET', 'POST'])
+def product_list(request):
+    if request.method == 'GET':
         queryset = Product.objects.select_related('collection').all()[:10]
         serializer = ProductSerializer(queryset, many=True, context={'request': request})
         return Response(serializer.data)
 
-    def post(self, request):
+    elif request.method == 'POST':
         serializer = ProductSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
+        serializer.validated_data
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
-class ProductDetail(APIView):
-    def get(self, request, id):
-        product = get_object_or_404(Product, pk=id)
+@api_view(['GET', 'PUT', 'DELETE'])
+def product_detail(request, id):
+    product = get_object_or_404(Product, pk=id)
+    if request.method == 'GET':
         serializer = ProductSerializer(product)
         return Response(serializer.data)
 
-    def put(self, request, id):
-        product = get_object_or_404(Product, pk=id)
+    elif request.method == 'PUT':
         serializer = ProductSerializer(product, data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data)
 
-    def delete(self, request, id):
-        product = get_object_or_404(Product, pk=id)
+    elif request.method == 'DELETE':
         if product.orderitems.count() > 0:
             return Response({'error': 'Product cannot be deleted because it is associated with an order item.'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
-
         product.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -72,41 +71,41 @@ def collection_detail(request, pk):
     elif request.method == 'DELETE':
         if collection.products.count() > 0:
             return Response({'error': 'Collection cannot be deleted because it includes one or more products.'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
-
         collection.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
-class CutomerList(APIView):
-    def get(self, request):
+@api_view(['GET', 'POST'])
+def customer_list(request):
+    if request.method == 'GET':
         queryset = Customer.objects.all()
         serializer = CustomerSerializer(queryset, many=True)
         return Response(serializer.data)
 
-    def post(self, request):
+    elif request.method == 'POST':
         serializer = CustomerSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
-class CustomerDetail(APIView):
-    def get(self, request, id):
-        customer = get_object_or_404(Customer, pk=id)
+@api_view(['GET', 'PUT', 'DELETE'])
+def customer_detail(request, id):
+    customer = get_object_or_404(Customer, pk=id)
+
+    if request.method == 'GET':
         serializer = CustomerSerializer(customer)
         return Response(serializer.data)
 
-    def put(self, request, id):
-        customer = get_object_or_404(Customer, pk=id)
+    elif request.method == 'PUT':
         serializer = CustomerSerializer(customer, data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save()
         return Response(serializer.data)
 
-    def delete(self, request, id):
-        customer = get_object_or_404(Customer, pk=id)
+    elif request.method == 'DELETE':
         if customer.order_set.count() > 0:
-            return Response({'error': 'Customer cannot be deleted because customer has one or more orders.'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
+            return Response({'error': 'Customer cannot be deleted because a order is associated with the customer.'}, status=status.HTTP_405_METHOD_NOT_ALLOWED)
 
         customer.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
